+++ asilib/asi/__init__.py
@@ -1,7 +1,7 @@
 from .rego import rego, rego_info, rego_skymap
-from .themis import themis, themis_info, themis_skymap
+from .themis import themis, themis_info, themis_skymap, themis_available, plot_themis_available
 from .trex import trex_nir, trex_nir_skymap, trex_nir_info
-from .trex import trex_rgb, trex_rgb_info, trex_rgb_skymap
+from .trex import trex_rgb, trex_rgb_info, trex_rgb_skymap, trex_rgb_available, plot_trex_rgb_available
 from .psa_project import psa_project, psa_project_info, psa_project_skymap, psa_project_lamp
 from .mango import mango, mango_info
 
@@ -9,6 +9,8 @@ __all__ = [
     'themis', 
     'themis_info', 
     'themis_skymap',
+    'themis_available',
+    'plot_themis_available',
     'rego',
     'rego_info', 
     'rego_skymap',
@@ -18,6 +20,8 @@ __all__ = [
     'trex_nir', 
     'trex_nir_skymap', 
     'trex_nir_info',
+    'trex_rgb_available', 
+    'plot_trex_rgb_available',
     'mango',
     'mango_info',
     'psa_project',