+++ asilib/imagers.py
@@ -65,7 +65,39 @@ class Imagers:
                 self.start_time = min(start_times)
                 self.end_time = max(end_times)
             self.n_times = int(np.ceil((self.end_time - self.start_time).total_seconds() / self.min_cadence))
+
+        self.lon_bounds, self.lat_bounds = self.spatial_extent()
         return
+    
+    def spatial_extent(self, min_elevation: float = 10) -> Tuple[Tuple[float, float], Tuple[float, float]]:
+        """
+        Calculate the longitude and latitude bounds of the mosaic.
+
+        Returns
+        -------
+        Tuple[float, float]
+            Longitude bounds.
+        Tuple[float, float]
+            Latitude bounds.
+        """
+        lon_bounds = [np.nan, np.nan]
+        lat_bounds = [np.nan, np.nan]
+        for imager in self.imagers:
+            _skymap_cleaner = Skymap_Cleaner(
+                imager.skymap['lon'], 
+                imager.skymap['lat'], 
+                imager.skymap['el'], 
+            )
+            _skymap_cleaner.mask_elevation(min_elevation)
+
+            _imager_lon_bounds = np.nanmin(_skymap_cleaner._lon_grid), np.nanmax(_skymap_cleaner._lon_grid)
+            _imager_lat_bounds = np.nanmin(_skymap_cleaner._lat_grid), np.nanmax(_skymap_cleaner._lat_grid)
+
+            lon_bounds[0] = np.nanmin([lon_bounds[0], _imager_lon_bounds[0]])
+            lon_bounds[1] = np.nanmax([lon_bounds[1], _imager_lon_bounds[1]])
+            lat_bounds[0] = np.nanmin([lat_bounds[0], _imager_lat_bounds[0]])
+            lat_bounds[1] = np.nanmax([lat_bounds[1], _imager_lat_bounds[1]])
+        return tuple(lon_bounds), tuple(lat_bounds)
     
     def plot_fisheye(self, ax:Tuple[plt.Axes], **kwargs):
         """
@@ -515,7 +547,7 @@ class Imagers:
         self.imagers[0]._create_animation(image_paths, movie_save_path, ffmpeg_params, overwrite)
         return
     
-    def map_eq(self, b_model: Callable='IGRF') -> Tuple[np.ndarray, np.ndarray]:
+    def map_eq(self, b_model: Callable='IGRF', normalize_intensities:bool=False, min_elevation:float=10) -> Tuple[np.ndarray, np.ndarray]:
         """
         Map an auroral image to the magnetic equator using IGRF or a user-defined magnetic field model.
 
@@ -525,6 +557,10 @@ class Imagers:
             A function with the (time, lla) arguments where time is a single time stamp, 
             and lla is one (lat, lon, alt) point. If ``b_model='IGRF'``, IRBEM's IGRF model is used
             in the default GSM coordinates.
+        normalize_intensities: bool
+            If True, the B&W pixel intensities are normalized to 0-1.
+        min_elevation: float
+            The minimum elevation angle for pixels to be included in the mapping.
 
         Returns
         -------
@@ -545,7 +581,10 @@ class Imagers:
                     'or supply a custom b_model.'
                     )
             
-        lats_lons, intensities = self.get_points()
+        lats_lons, intensities = self.get_points(
+            normalize_intensities=normalize_intensities,
+            min_elevation=min_elevation,
+            )
         equator_sm = np.zeros((lats_lons[:, 0].shape[0], 3))
 
         _progressbar = asilib.utils.progressbar(
@@ -562,10 +601,12 @@ class Imagers:
             self, 
             ax:plt.Axes=None, 
             b_model:Callable="IGRF",
+            min_elevation:float=10,
             max_valid_grid_distance:float=0.03,
             x_grid:np.ndarray=None, 
             y_grid:np.ndarray=None, 
             color_bounds: List[float] = None,
+            normalize_intensities: bool = False,
             color_norm: str = None,
             color_map: str = None,
             pcolormesh_kwargs:dict={},
@@ -582,6 +623,8 @@ class Imagers:
             A function with the (time, lla) arguments where time is a single time stamp, 
             and lla is one (lat, lon, alt) point. If ``b_model='IGRF'``, IRBEM's IGRF model is used
             in the default GSM coordinates.
+        min_elevation: float
+            The minimum elevation angle for pixels to be included in the mapping.
         max_valid_grid_distance: float
             The maximum distance from the mapped point to the grid point. If the distance is greater
             than this value, the mapped grid point will be masked as NaN. The distance is in units of Re.
@@ -590,6 +633,8 @@ class Imagers:
             -12 < x < 1.1 and -5 < y < 5 is generated.
         color_bounds: List[float]
             The ASI image color bounds.
+        normalize_intensities: bool
+            If True, the B&W pixel intensities are normalized to 0-1 before plotting
         color_map: str
             The matplotlib colormap to use. See `matplotlib colormaps <https://matplotlib.org/stable/tutorials/colors/colormaps.html>`_
             for supported colormaps.
@@ -605,7 +650,7 @@ class Imagers:
         if (x_grid is None) and (y_grid is None):
             x_grid, y_grid = np.meshgrid(np.linspace(-12, 1.1, num=1000), np.linspace(-5, 5, num=1001))
 
-        equator_sm, intensities = self.map_eq(b_model=b_model)
+        equator_sm, intensities = self.map_eq(b_model=b_model, min_elevation=min_elevation)
 
         gridded_eq_data = scipy.interpolate.griddata(
             equator_sm[:, :2], 
@@ -616,9 +661,16 @@ class Imagers:
         # https://stackoverflow.com/a/31189177
         tree = cKDTree(equator_sm[:, :2])
         dists, _ = tree.query(np.stack((x_grid, y_grid), axis=-1))
-        gridded_eq_data[dists > max_valid_grid_distance, :] = np.nan  # Mask any gridded point > 0.1 Re from the mapped point as NaN
+        gridded_eq_data[dists > max_valid_grid_distance, ...] = np.nan  # Mask any gridded point > 0.1 Re from the mapped point as NaN
 
-        color_map, color_norm = self.imagers[0]._plot_params(gridded_eq_data, color_bounds, color_map, color_norm)
+        if not normalize_intensities:
+            _color_map, color_norm = self.imagers[0]._plot_params(gridded_eq_data, color_bounds, color_map, color_norm)
+        else:
+            _color_map = None
+            color_norm = matplotlib.colors.Normalize(vmin=0, vmax=1)
+        
+        if color_map is None:
+            color_map = _color_map
 
         vmin, vmax = self.imagers[0].plot_settings['color_bounds']
         if len(self.imagers[0].meta['resolution']) == 3:
@@ -758,7 +810,7 @@ class Imagers:
             yield guide_time, _asi_times, _asi_images
         return
 
-    def get_points(self, min_elevation:float=10)->Tuple[np.ndarray, np.ndarray]:
+    def get_points(self, min_elevation:float=10, normalize_intensities:bool=False)->Tuple[np.ndarray, np.ndarray]:
         """
         Get pixel intensities in each (lat, lon) grid point.
 
@@ -766,6 +818,8 @@ class Imagers:
         ----------
         min_elevation: float
             Only return pixel intensities above min_elevation.
+        normalize_intensities: bool
+            If True, the returned pixel intensities are normalized to the range [0, 1] using the imager's color bounds.
 
         Returns
         -------
@@ -872,6 +926,11 @@ class Imagers:
             lat_grid = _masked_lat_map[_valid_idx[0], _valid_idx[1]]
             lon_grid = _masked_lon_map[_valid_idx[0], _valid_idx[1]]
             intensity = _imager.data.image[_valid_idx[0], _valid_idx[1], ...]
+
+            if (len(self.imagers[0].meta['resolution']) == 2) and normalize_intensities: # B&W images only.
+                old_min, old_max = _imager.get_color_bounds()
+                new_min, new_max = 0, 1
+                intensity = (intensity - old_min) / (old_max - old_min) * (new_max - new_min) + new_min
 
             # Concatenate joins arrays along an existing axis, while stack joins arrays
             # along a new axis. 
@@ -1097,6 +1156,25 @@ class Imagers:
                     overlap_mask[ij_overlap_pixels] = True
                     overlap_pixels[self_key][other_key] = overlap_mask
         return overlap_pixels
+    
+    def __getitem__(self, _slice):
+        """
+        Slice all Imagers objects by time.
+
+        Parameters
+        ----------
+        _slice: str, pd.Timestamp, datetime.datetime, or list.
+            The time(s) to slice an Imager object. Can type of slice can be either
+            1) [start_time:end_time], or 2) just time.
+
+        Yields
+        ------
+        asilib.Imagers:
+            A sliced version of the Imagers.
+        """
+        _imagers = [_imager[_slice] for _imager in self.imagers]
+        cls = type(self)
+        return cls(_imagers)
     
     def __str__(self):
         names = [f'{_img.meta["array"]}-{_img.meta["location"]}' for _img in self.imagers]